services/veo_client.py
import os
import time
from typing import Dict, Optional
import google.generativeai as genai

# Configure Google AI with API key
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

def generate_video_with_veo3(prompt: str, length_seconds: int = 30, aspect_ratio: str = "9:16") -> Dict:
    """
    Generate video using Google's Veo 3 model via Gemini API.
    
    Args:
        prompt (str): Video generation prompt
        length_seconds (int): Video length in seconds (5-60)
        aspect_ratio (str): Aspect ratio ("16:9", "9:16", or "1:1")
    
    Returns:
        dict: Video generation result with status and video_url or error
    """
    try:
        # Create the model
        model = genai.GenerativeModel("gemini-2.0-flash-exp")
        
        # Format the prompt for video generation
        video_prompt = f"""Generate a {length_seconds}-second vertical video ({aspect_ratio} aspect ratio) with this concept:

{prompt}

Video requirements:
- High quality cinematic visuals
- Smooth camera movements
- Professional lighting
- Clean composition
- Duration: exactly {length_seconds} seconds
- Aspect ratio: {aspect_ratio}"""

        print(f"Generating video with Veo 3...")
        print(f"Prompt: {video_prompt[:200]}...")
        
        # Generate video (this is a placeholder - actual Veo 3 integration may differ)
        response = model.generate_content([
            "Create a video based on this prompt:",
            video_prompt
        ])
        
        # For now, return a success status with the generated content
        # In actual implementation, this would return video file or URL
        return {
            "status": "success",
            "message": "Video generation initiated successfully",
            "video_url": None,  # Would contain actual video URL
            "generation_id": f"veo3_{int(time.time())}",
            "estimated_completion": f"{length_seconds * 2} seconds",
            "prompt_used": video_prompt,
            "response": str(response.text)[:500] if response.text else "Generated successfully"
        }
        
    except Exception as e:
        print(f"Veo 3 generation error: {str(e)}")
        return {
            "status": "error",
            "error": str(e),
            "message": "Failed to generate video with Veo 3"
        }

def check_video_status(generation_id: str) -> Dict:
    """
    Check the status of a video generation job.
    
    Args:
        generation_id (str): The generation job ID
        
    Returns:
        dict: Status information
    """
    try:
        # This would check the actual generation status
        # For now, return a mock status
        return {
            "status": "processing",
            "progress": 75,
            "estimated_remaining": "30 seconds",
            "generation_id": generation_id
        }
        
    except Exception as e:
        return {
            "status": "error",
            "error": str(e),
            "generation_id": generation_id
        }

def build_veo3_prompt_from_script(script_json: Dict, brand_voice: str, cta: str) -> str:
    """
    Convert the generated script into a Veo 3 video generation prompt.
    
    Args:
        script_json (dict): Generated script with voiceover_lines, visuals, etc.
        brand_voice (str): Brand voice tone
        cta (str): Call to action
        
    Returns:
        str: Optimized video generation prompt
    """
    voiceover_lines = script_json.get("voiceover_lines", [])
    visuals = script_json.get("visuals", [])
    
    # Build narrative from voiceover
    narrative = " ".join(voiceover_lines)
    
    # Build visual direction
    visual_direction = ", ".join(visuals) if visuals else "professional business setting"
    
    # Create comprehensive prompt
    prompt = f"""Create a compelling {brand_voice} marketing video with this narrative:

NARRATIVE: {narrative}

VISUAL STYLE: {visual_direction}

REQUIREMENTS:
- {brand_voice} tone throughout
- Professional cinematography with smooth transitions
- End with clear call-to-action: "{cta}"
- Modern, engaging visual effects
- High production value
- Clear, readable text overlays when needed

MOOD: Confident, professional, engaging"""

    return prompt