pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "openai>=1.40.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "oauthlib>=3.3.1",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
    "stripe>=12.4.0",
    "beautifulsoup4>=4.13.4",
    "pillow>=11.3.0",
    "anthropic>=0.61.0",
    "bleach>=6.2.0",
    "markupsafe>=3.0.2",
    "sqlalchemy>=2.0.42",
    "google-genai>=1.29.0",
    "pytrends>=4.9.2",
    "scikit-learn>=1.7.1",
    "nltk>=3.9.1",
    "readability-lxml>=0.8.4.1",
]
